Added Integration Tests For The Authentication Controller Of The Gateway API
I added all the necessary integration tests for the authentication controller of the gateway api. For that purpose certain changes were made in the the authentication controller when issues were found. Additionally I added Api key authentication and bandwidth limit to the email microservice. A small change was made to the authentication controller with it now invalidating previous access token(check on the email of the access token) upon email change. One important change is that health checks were added to all 4 microservices. The reason for this is state control. In some endpoints of the gateway api more than one microservice is called. If a system error appears in one of the microservices after some state change has occured in a previous microservice then I would have to make some sort of rollback. Since rollback logic is pretty complex and I have not implemented it yet the health checks check before any state changes occur that all the microservices that the endpoint depends upon are online and healthy(database, smtp server etc are ok). This health checks in combination with the retry logic that I have implemented should make sure that an invalid state should not happen. One small thing I need to add is that I could not make the external login work with the gateway api, which means that the client will have to bypass the gateway api for this one procedure. I am not happy with this and I might revisit it in the future, but as it stands I have not added the necessary logic for this endpoint.

+++ EshopApp.GatewayAPI/Controllers/GatewayAuthenticationController.cs
@@ -1,7 +1,7 @@
-﻿using EshopApp.GatewayAPI.Models.RequestModels;
+﻿using EshopApp.GatewayAPI.Models;
+using EshopApp.GatewayAPI.Models.RequestModels;
 using EshopApp.GatewayAPI.Models.ServiceRequestModels;
 using EshopApp.GatewayAPI.Models.ServiceResponseModels;
-using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.RateLimiting;
 using System.IdentityModel.Tokens.Jwt;
@@ -17,6 +17,7 @@ namespace EshopApp.GatewayAPI.Controllers;
 [Route("api/[controller]")]
 public class GatewayAuthenticationController : ControllerBase
 {
+    //the general idea with redirectUrl is that it can contain a returnUrl that the frontend handler(the endpoint that is specified as the redirectUrl) can use to redirect the user too after its own processing
     private readonly HttpClient authHttpClient;
     private readonly HttpClient emailHttpClient;
     private readonly HttpClient dataHttpClient;
@@ -37,7 +38,7 @@ public class GatewayAuthenticationController : ControllerBase
         SetDefaultHeadersForClient(true, authHttpClient, _configuration["AuthApiKey"]!, _configuration["AuthRateLimitingBypassCode"]!);
         HttpResponseMessage? response = await authHttpClient.GetAsync("Authentication/GetUserByAccessToken");
 
-        //validate that changing the password has worked
+        //validate that getting the user has worked
         int retries = 3;
         while ((int)response.StatusCode >= 500)
         {
@@ -48,14 +49,19 @@ public class GatewayAuthenticationController : ControllerBase
             retries--;
         }
 
-        if ((int)response.StatusCode >= 400 && (int)response.StatusCode < 500) //you can get BadRequest from validationErrors
+        if ((int)response.StatusCode >= 400 && (int)response.StatusCode < 500)
             return await CommonValidationForRequestClientErrorCodesAsync(response);
 
         string? responseBody = await response.Content.ReadAsStringAsync();
-        JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("AccessToken", out string? accessToken);
-        return Ok();
+        GatewayAppUser? appUser = JsonSerializer.Deserialize<GatewayAppUser>(responseBody, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
+        return Ok(appUser);
     }
 
+    //The client Url might other than the handler endpoint might contain as a query parameter something like returnUrl, which can be used
+    //by the handler endpoint to redirect back to the whole signup started from. So if the signup started from signup then redirect to home
+    //if it started from some part inside the application(due to need authentication) then redirect there instead of the default, which is home
+    //I suppose I will leave that for the front end implementation to consider. The only thing that is needed 100% is the domain:port/endpoint to be sent
+    //but if the frontend wants it can be domain:port/endpoint?returnUrl=returnUrl
     [HttpPost("SignUp")]
     public async Task<IActionResult> SignUp([FromBody] GatewayApiSignUpRequestModel signUpModel)
     {
@@ -69,6 +75,10 @@ public class GatewayAuthenticationController : ControllerBase
             //remove the trailing slash from the client url
             if (signUpModel.ClientUrl!.EndsWith("/"))
                 signUpModel.ClientUrl = signUpModel.ClientUrl.Substring(0, signUpModel.ClientUrl.Length - 1);
+
+            //start by doing healthchecks for the endpoints this is calling
+            if (!await CheckIfMicroservicesFullyOnlineAsync(new List<HttpClient>() { authHttpClient, dataHttpClient, emailHttpClient }))
+                return StatusCode(503, new { ErrorMessage = "OneOrMoreMicroservicesAreUnavailable" });
 
             //sign up user
             GatewayApiSignUpServiceRequestModel signUpRequestModel = new GatewayApiSignUpServiceRequestModel();
@@ -115,8 +125,9 @@ public class GatewayAuthenticationController : ControllerBase
                 return await CommonValidationForRequestClientErrorCodesAsync(response);
 
             //send confirmation email
-            string message = "Click on the following link to confirm your email: ";
-            string link = $"{signUpModel.ClientUrl}?userId={signupResponseModel!.UserId}&token={WebUtility.UrlEncode(signupResponseModel.ConfirmationToken)}";
+            string message = "Click on the following link to confirm your email:";
+            string link = $"{_configuration["AuthApiBaseUrl"]}Authentication/ConfirmEmail?" +
+                $"userId={signupResponseModel!.UserId}&confirmEmailToken={WebUtility.UrlEncode(signupResponseModel.ConfirmationToken)}&redirectUrl={WebUtility.UrlEncode(signUpModel.ClientUrl)}";
             string? confirmationLink = $"{message} {link}";
             var apiSendEmailModel = new Dictionary<string, string>
             {
@@ -167,7 +178,7 @@ public class GatewayAuthenticationController : ControllerBase
 
             //return access accessToken
             string? responseBody = await response.Content.ReadAsStringAsync();
-            JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("AccessToken", out string? accessToken);
+            JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("accessToken", out string? accessToken);
             return Ok(new { AccessToken = accessToken });
         }
         catch
@@ -176,49 +187,8 @@ public class GatewayAuthenticationController : ControllerBase
         }
     }
 
-    [HttpPost("ExternalSignIn")]
-    public async Task<IActionResult> ExternalSignIn([FromBody] GatewayApiExternalSignInRequestModel externalSignInModel)
-    {
-        try
-        {
-            //externally start the sign in of the user
-            SetDefaultHeadersForClient(false, authHttpClient, _configuration["AuthApiKey"]!, _configuration["AuthRateLimitingBypassCode"]!);
-            HttpResponseMessage? response = await authHttpClient.PostAsJsonAsync("Authentication/ExternalSignIn",
-                new { ReturnUrl = externalSignInModel.ReturnUrl, IdentityProviderName = externalSignInModel.IdentityProviderName });
-
-            //validation that sign in has worked
-            int retries = 3;
-            while ((int)response.StatusCode >= 500)
-            {
-                if (retries == 0)
-                    return StatusCode(500, "Internal Server Error");
-
-                response = await authHttpClient.PostAsJsonAsync("Authentication/ExternalSignIn", await authHttpClient.PostAsJsonAsync("Authentication/ExternalSignIn",
-                    new { ReturnUrl = externalSignInModel.ReturnUrl, IdentityProviderName = externalSignInModel.IdentityProviderName }));
-                retries--;
-            }
-
-            if ((int)response.StatusCode >= 400 && (int)response.StatusCode < 500)
-                return await CommonValidationForRequestClientErrorCodesAsync(response);
-
-            //return the challenge result to the client
-            string? responseBody = await response.Content.ReadAsStringAsync();
-            var challengeDetails = JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody);
-            string? identityProviderName = challengeDetails?["IdentityProviderName"];
-            string? redirectUri = challengeDetails?["RedirectUri"];
-
-            var authProperties = new AuthenticationProperties
-            {
-                RedirectUri = redirectUri
-            };
-            return new ChallengeResult(identityProviderName!, authProperties);
-        }
-        catch
-        {
-            return StatusCode(500, "Internal Server Error");
-        }
-    }
-
+    //the clientUrl here contains just the front end handler since the reset password endpoint that is need to be called after this requires more information from the front end
+    //so the clientUrl is something like this: domain:port/endpoint with this endpoint appending userId and password reset token(so the front end handler needs to include those)
     [HttpPost("ForgotPassword")]
     public async Task<IActionResult> ForgotPassword([FromBody] GatewayApiForgotPasswordRequestModel forgotPasswordModel)
     {
@@ -231,6 +201,10 @@ public class GatewayAuthenticationController : ControllerBase
             //remove the trailing slash from the client url
             if (forgotPasswordModel.ClientUrl!.EndsWith("/"))
                 forgotPasswordModel.ClientUrl = forgotPasswordModel.ClientUrl.Substring(0, forgotPasswordModel.ClientUrl.Length - 1);
+
+            //start by doing healthchecks for the endpoints this is calling
+            if (!await CheckIfMicroservicesFullyOnlineAsync(new List<HttpClient>() { authHttpClient, emailHttpClient }))
+                return StatusCode(503, new { ErrorMessage = "OneOrMoreMicroservicesAreUnavailable" });
 
             //request signin in the user
             SetDefaultHeadersForClient(false, authHttpClient, _configuration["AuthApiKey"]!, _configuration["AuthRateLimitingBypassCode"]!);
@@ -253,7 +227,7 @@ public class GatewayAuthenticationController : ControllerBase
             string? responseBody = await response.Content.ReadAsStringAsync();
             GatewayApiForgotPasswordServiceResponseModel? forgotPasswordResponseModel = JsonSerializer.Deserialize<GatewayApiForgotPasswordServiceResponseModel>(responseBody, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
 
-            string message = "Click on the following link to change your account's password: ";
+            string message = "Click on the following link to change your account's password:";
             string link = $"{forgotPasswordModel.ClientUrl}?userId={forgotPasswordResponseModel!.UserId}&token={WebUtility.UrlEncode(forgotPasswordResponseModel.Token)}";
             string? confirmationLink = $"{message} {link}";
             var apiSendEmailModel = new Dictionary<string, string>
@@ -302,7 +276,7 @@ public class GatewayAuthenticationController : ControllerBase
                 return await CommonValidationForRequestClientErrorCodesAsync(response);
 
             string? responseBody = await response.Content.ReadAsStringAsync();
-            JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("AccessToken", out string? accessToken);
+            JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("accessToken", out string? accessToken);
             return Ok(new { AccessToken = accessToken });
         }
         catch
@@ -343,6 +317,9 @@ public class GatewayAuthenticationController : ControllerBase
         }
     }
 
+    //the clientUrl in this works the same way as in the signup, which means that the format is something along those lines: domain:port/endpoint now if the front end wants it can be
+    //domain:port/endpoint?returnUrl=returnUrl with the front end handler using the returnUrl to redirect the user after everything is done to where this whole process started from
+    //(in this case it is probably pointless since it should return them to home no matter what, but who knows)
     [HttpPost("RequestChangeAccountEmail")]
     public async Task<IActionResult> RequestChangeAccountEmail([FromBody] GatewayApiChangeEmailRequestModel changeEmailModel)
     {
@@ -355,6 +332,10 @@ public class GatewayAuthenticationController : ControllerBase
             //remove the trailing slash from the client url
             if (changeEmailModel.ClientUrl!.EndsWith("/"))
                 changeEmailModel.ClientUrl = changeEmailModel.ClientUrl.Substring(0, changeEmailModel.ClientUrl.Length - 1);
+
+            //start by doing healthchecks for the endpoints this is calling
+            if (!await CheckIfMicroservicesFullyOnlineAsync(new List<HttpClient>() { authHttpClient, emailHttpClient }))
+                return StatusCode(503, new { ErrorMessage = "OneOrMoreMicroservicesAreUnavailable" });
 
             //request to change the email of the user
             string accessToken = SetDefaultHeadersForClient(true, authHttpClient, _configuration["AuthApiKey"]!, _configuration["AuthRateLimitingBypassCode"]!)!;
@@ -377,14 +358,15 @@ public class GatewayAuthenticationController : ControllerBase
             //after this point the token is certainly valid
             //send the change email link to the user's new email
             string? responseBody = await response.Content.ReadAsStringAsync();
-            JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("ChangeEmailToken", out string? changeEmailToken);
+            JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody)!.TryGetValue("changeEmailToken", out string? changeEmailToken);
 
             var handler = new JwtSecurityTokenHandler();
             var jwtToken = handler.ReadJwtToken(accessToken);
             string userId = jwtToken.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value!;
 
-            string message = "Click on the following link to confirm your account's new email: ";
-            string? link = $"{changeEmailModel.ClientUrl}?userId={userId}&newEmail={changeEmailModel.NewEmail}&token={WebUtility.UrlEncode(changeEmailToken)}";
+            string message = "Click on the following link to confirm your account's new email:";
+            string link = $"{_configuration["AuthApiBaseUrl"]}Authentication/ConfirmChangeEmail" +
+                $"?userId={userId}&newEmail={changeEmailModel.NewEmail}&changeEmailToken={WebUtility.UrlEncode(changeEmailToken)}&redirectUrl={WebUtility.UrlEncode(changeEmailModel.ClientUrl)}";
             string? confirmationLink = $"{message} {link}";
             var apiSendEmailModel = new Dictionary<string, string>
             {
@@ -411,6 +393,10 @@ public class GatewayAuthenticationController : ControllerBase
     {
         try
         {
+            //start by doing healthchecks for the endpoints this is calling
+            if (!await CheckIfMicroservicesFullyOnlineAsync(new List<HttpClient>() { authHttpClient, dataHttpClient, emailHttpClient }))
+                return StatusCode(503, new { ErrorMessage = "OneOrMoreMicroservicesAreUnavailable" });
+
             //request to delete the account of the user
             string accessToken = SetDefaultHeadersForClient(true, authHttpClient, _configuration["AuthApiKey"]!, _configuration["AuthRateLimitingBypassCode"]!)!;
             HttpResponseMessage? response = await authHttpClient.DeleteAsync("Authentication/DeleteAccount");
@@ -454,7 +440,7 @@ public class GatewayAuthenticationController : ControllerBase
                 return await CommonValidationForRequestClientErrorCodesAsync(response);
 
             //request to remove all the user coupons
-            response = await dataHttpClient.DeleteAsync($"Cart/UserId/{userId}");
+            response = await dataHttpClient.DeleteAsync($"Coupon/RemoveAllUserCoupons/userId/{userId}");
 
             //validate the deletion of the usercoupons
             retries = 3;
@@ -470,7 +456,7 @@ public class GatewayAuthenticationController : ControllerBase
             if ((int)response.StatusCode >= 400 && (int)response.StatusCode < 500)
                 return await CommonValidationForRequestClientErrorCodesAsync(response);
 
-            //send a request to the user that their account has been deleted
+            //send an email to the user to notify them that their account has been deleted
             var apiSendEmailModel = new Dictionary<string, string>
             {
                 { "receiver", email },
@@ -524,9 +510,28 @@ public class GatewayAuthenticationController : ControllerBase
     private async Task<IActionResult> CommonValidationForRequestClientErrorCodesAsync(HttpResponseMessage response)
     {
         string responseBody = await response.Content.ReadAsStringAsync();
-        var keyValue = JsonSerializer.Deserialize<Dictionary<string, string>>(responseBody);
-        keyValue!.TryGetValue("ErrorMessage", out string? errorMessage);
-        keyValue!.TryGetValue("Errors", out var errors);
+        //in the case there is no body
+        if (string.IsNullOrEmpty(responseBody))
+        {
+            if (response.StatusCode == HttpStatusCode.Unauthorized)
+                return Unauthorized();
+            else if (response.StatusCode == HttpStatusCode.Forbidden)
+                return StatusCode(StatusCodes.Status403Forbidden);
+            else if (response.StatusCode == HttpStatusCode.BadRequest)
+                return BadRequest();
+            else if (response.StatusCode == HttpStatusCode.NotFound)
+                return NotFound();
+            else if (response.StatusCode == HttpStatusCode.MethodNotAllowed)
+                return NotFound();
+
+            return BadRequest(); //this will probably never happen
+        }
+
+        //otherwise
+        var keyValue = JsonSerializer.Deserialize<Dictionary<string, object>>(responseBody);
+        keyValue!.TryGetValue("errorMessage", out object? errorMessageObject);
+        string? errorMessage = errorMessageObject?.ToString() ?? null;
+        keyValue!.TryGetValue("errors", out var errors);
 
         if (response.StatusCode == HttpStatusCode.Unauthorized && errorMessage is not null)
             return Unauthorized(new { ErrorMessage = errorMessage });
@@ -547,7 +552,7 @@ public class GatewayAuthenticationController : ControllerBase
         else if (response.StatusCode == HttpStatusCode.MethodNotAllowed)
             return StatusCode(StatusCodes.Status405MethodNotAllowed);
 
-        return BadRequest();
+        return BadRequest(); //this will probably never happen
     }
 
     private bool CheckIfUrlIsTrusted(string redirectUrl)
@@ -563,5 +568,28 @@ public class GatewayAuthenticationController : ControllerBase
         }
 
         return false;
+    }
+
+    private async Task<bool> CheckIfMicroservicesFullyOnlineAsync(List<HttpClient> httpClients)
+    {
+        try
+        {
+            if (httpClients is null || !httpClients.Any())
+                return true;
+
+            foreach (HttpClient client in httpClients)
+            {
+                var healthResponse = await authHttpClient.GetAsync("Health");
+                if (healthResponse.StatusCode != HttpStatusCode.OK)
+                    return false;
+            }
+
+            return true;
+        }
+        //the exception can happen if one of the microservices are not online. In this case just return false
+        catch
+        {
+            return false;
+        }
     }
 }
